big update to make the tests work
THIS IS PROBABLY A BREAKING CHANGE, SO GO SLOW & CAREFUL
WHEN YOU APPLY TO PROD.
Models and migrations were re-created from scratch such that
they work properly.  Previously, most models had managed=False
in them, which meant that changes had to be applied manually
to the database.  This was fine until it was time to add tests.
You'll want to migrate --fake on live data after truncating the
django_migrations table in the database.
Then watch it carefully.

The tests do not currently pass (need to add a silly amount of
tests data first) but they do run.